requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.0
SQLAlchemy>=2.0.36
Werkzeug==3.0.1
requests==2.31.0
twilio==8.10.0
python-dateutil==2.8.2

# Optional: PostgreSQL support (uncomment if using PostgreSQL)
# psycopg2-binary==2.9.9

